filters changes

+++ src/main/java/dashboard/dto/PageDTO.java
@@ -15,7 +15,7 @@ public class PageDTO {
 	
 	int current_page;
 	int items_on_page;
-	int total_count;
+	Long total_count;
 	OrderPageDTO orderPage;
 
 }
